src/requirements.txt
openai 
python-dotenv
requests 
markdownify
langchain
supabase
tqdm